+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Tabs no longer focus on initialization
 - mwc-list-item ripple color will now change based off of --mdc-ripple-color on initialization
 - Fix issue where textfield would throw an error when fed a non-string value
+- list selected item will update if selected item is disconnected
 
 ### Added